+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Erl2ex is currently pre-alpha software. Expect significant backwards-incompatibl
 ## v0.0.2 (????-??-??)
 
 *   Much better reporting of parse and conversion errors
+*   Support for custom and remote types
 
 ## v0.0.1 (2015-12-28)